% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petroOne-package.r
\docType{data}
\name{custom_stopwords}
\alias{custom_stopwords}
\title{Default custom stop words}
\format{An object of class \code{character} of length 133.}
\usage{
custom_stopwords
}
\description{
This is a minimal dataset of custom stopwords.
Tou can supply your own stopwords by editing the file stopwords.txt and
then importing them. But this is a basic way to start and eliminate
common words from the paper titles during clasification.
Dataset: stopwords.rda
Source:  stopwords.txt
Class:   char vector
}
\keyword{datasets}
