% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{get_papers_count}
\alias{get_papers_count}
\title{Number of paper for a given query}
\usage{
get_papers_count(url)
}
\arguments{
\item{url}{char a query URL for OnePetro}
}
\description{
Obtains the number of papers being queried by the URL
}
