% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{make_search_url}
\alias{make_search_url}
\title{Make a search URL for OnePetro}
\usage{
make_search_url(query = NULL, start = NULL, from_year = NULL,
  peer_reviewed = NULL, published_between = NULL, rows = NULL,
  to_year = NULL, dc_type = NULL, how = "any")
}
\description{
Create a URL that works in OnePetro website
}
