% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_by_dates}
\alias{summary_by_dates}
\title{Get summary dataframe by year}
\usage{
summary_by_dates(result)
}
\arguments{
\item{result}{a OnePetro page with results}
}
\description{
Year that the paper was published
}
