% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_by_doctype}
\alias{summary_by_doctype}
\title{Get summary dataframe by document type}
\usage{
summary_by_doctype(result)
}
\arguments{
\item{result}{a OnePetro page with results}
}
\description{
Document types are: conference-paper, journal-paper, etc.
}
