% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_by_publications}
\alias{summary_by_publications}
\title{Get summary dataframe by publication}
\usage{
summary_by_publications(result)
}
\arguments{
\item{result}{a OnePetro page with results}
}
\description{
Publications could be World Petroleum Congress, Annual Technical
Meeting, SPE Unconventional Reservoirs Conference, etc.
}
