% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_by_publisher}
\alias{summary_by_publisher}
\title{Get summary dataframe by publisher}
\usage{
summary_by_publisher(result)
}
\arguments{
\item{result}{a OnePetro page with results}
}
\description{
Summary dataframe by publisher: OTC, SPE, etc.
}
